Improving CLI and GUI (#163)
* Improving CLI and GUI

* Update http_client.py

* Update http_client.py

* Update http_client.py

* Update test_view.py

* Update test_view.py

* Update test_view.py

* Update view.py

+++ ui/src/main/java/com/intel/dai/ui/JsonConverter.java
@@ -30,7 +30,7 @@ public class JsonConverter {
                     rowData.add( resultsetinp.getArray(i));
                 }
                 else if(rasmetadata.getColumnType(i)== Types.BIGINT){
-                    rowData.add( resultsetinp.getInt(i));
+                    rowData.add( resultsetinp.getLong(i));
                 }
                 else if(rasmetadata.getColumnType(i)== Types.TINYINT){
                     rowData.add( resultsetinp.getInt(i));
